minor restructuring for clarity
