+++ internal/util/io.go
@@ -1,4 +1,4 @@
-package io
+package util
 
 import (
 	"fmt"
